backend/requirements.txt
pandas==2.0.3
numpy==1.24.3
boto3==1.28.38
pymongo==4.4.1
openai==1.3.7 
requests==2.31.0
flask-cors
python-dotenv==1.0.0
flask_pymongo
matplotlib==3.7.2  # For any potential data visualization
scikit-learn==1.3.0  # For any machine learning tasks
tqdm==4.66.1  # For progress bars
pytest==7.4.0  # For testing
certifi
colorama
pytz # pst time zones
tenacity


#python3 -m venv venv
#source venv/bin/activate  

#pip install -r requirements.txt
